feat(sim): match gamemd Drive Ship slope lifecycle
gamemd.exe evidence: Drive Process 0x004B0500 and Draw_Matrix 0x004AFF60; Ship Process 0x0069FC10 and Draw_Matrix 0x0069F670; FootClass::Unlimbo 0x004D7170; CDTimer Remaining 0x004B4D70; Drive/Ship Save 0x004AF800/0x0069EF10.

+++ src/render/unit_slope_transition_cache.rs
@@ -27,7 +27,7 @@ pub struct TransitionUnitSpriteKey {
     pub frame: u32,
     pub from_slope: u8,
     pub to_slope: u8,
-    pub phase_num: u8,
+    pub phase_num: i32,
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -207,7 +207,7 @@ impl VxlSlopeTransitionCache {
 mod tests {
     use super::*;
 
-    fn key(from_slope: u8, to_slope: u8, phase_num: u8) -> TransitionUnitSpriteKey {
+    fn key(from_slope: u8, to_slope: u8, phase_num: i32) -> TransitionUnitSpriteKey {
         TransitionUnitSpriteKey {
             type_id: "CMIN".to_string(),
             facing: 64,
